feat: restructure project layout, implement new landing, login, and register pages, and enhance routing with updated components
