+++ admin-proyect/src/pages/dashboard/Dashboard.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function Dashboard(){
+    return (
+        <div className="min-vh-100 bg-dark">
+            <div className="text-white p-4">
+                <h1>Dashboard principal</h1>
+                {/* Aquí irá el contenido del dashboard */}
+            </div>
+        </div>
+    );
+}
+
+export default Dashboard;
