+++ admin-proyect/src/pages/landing/README.md
@@ -1,0 +1,14 @@
+# LANDING
+---
+
+Este apartado nos servira para llamar la atencion del cliente, este sera sencillo pero concreto sin informacion de más que no sature al cliente.
+
+---
+
+### Imagen
+
+- No es el resultado final pero es una representacion grafica de como deberia de quedar.
+
+<p align="center">
+  <img src="https://i.postimg.cc/XYbTjZD2/Captura-de-pantalla-2025-04-30-153510.png" alt="Imagen">
+</p>
