+++ admin-proyect/src/pages/settings/Settings.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function Settings() {
+    return (
+        <div className="min-vh-100 bg-dark">
+            <div className="text-white p-4">
+                <h1>Configuración</h1>
+                {/* Aquí irá el contenido de configuración */}
+            </div>
+        </div>
+    );
+}
+
+export default Settings;
