+++ admin-proyect/src/routes/AppRoutes.jsx
@@ -1,7 +1,7 @@
 import { Routes, Route } from 'react-router-dom';
-import Landing from '../views/Landing';
-import Login from '../views/Login';
-import Register from '../views/Register';
+import Landing from '../pages/landing/Landing';
+import Login from '../pages/auth/login/Login';
+import Register from '../pages/auth/register/Register';
 
 export function AppRoutes() {
   return (
